src/requirements.txt
python-dotenv==1.0.1
python-telegram-bot==21.10
requests==2.32.3
schedule==1.2.2
